requirements.txt
gTTS==2.3.2
openai==0.27.6
playsound==1.3.0
python-dotenv==1.0.0
sounddevice==0.4.6
wavio==0.0.7
